Separates format of CSV results from file I/O
